restrucure for proper pip install
