+++ Cargo.toml
@@ -1,16 +1,17 @@
 [package]
 name = "tpx3_toolkit"
-version = "0.1.0"
+version = "0.2.0"
+authors = ["Brayden Freitas <andrewbrayden98@gmail.com>"]
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "rust_parse"
+name = "tpx3_toolkit"
 crate-type = ["cdylib"]
 
 [profile.release]
 #debug = true
 
 [dependencies]
-pyo3 = "0.18.1"
+pyo3 =  {version = "0.18.1", features = ["extension-module"]}
 numpy = "0.18"
