+++ pyproject.toml
@@ -4,13 +4,23 @@ build-backend = "maturin"
 
 [project]
 name = "tpx3_toolkit"
+description = "A package with tools for opening, viewing, and manipulating *.tpx3 output by the TimePix3 camera."
+readme = "README.md"
+authors = [
+    { name = "Brayden Freitas", email = "andrewbrayden98@gmail.com" },
+]
+version = "0.2.0"
 requires-python = ">=3.7"
+dependencies = [
+    "numpy >= 1.24.3",
+    "matplotlib >= 3.7.1",
+    "scipy >= 1.10.1",
+]
 classifiers = [
     "Programming Language :: Rust",
-    "Programming Language :: Python :: Implementation :: CPython",
-    "Programming Language :: Python :: Implementation :: PyPy",
+    "Programming Language :: Python :: 3",
 ]
 
-
 [tool.maturin]
-features = ["pyo3/extension-module"]
+module-name = "tpx3_toolkit.rust_parse"
+python-source = "python"
