--- setup.py
@@ -1,13 +1,0 @@
-from setuptools import setup
-
-setup(name="tpx3_toolkit",
-      version='0.1',
-      description='A package with tools for opening, viewing, and manipulating *.tpx3 output by the TimePix3 camera.',
-      author='Brayden Freitas',
-      author_email='andrewbrayden98@gmail.com',
-      packages=['tpx3_toolkit'],
-      install_requires=[
-        'numpy',
-        'scipy',
-        'matplotlib'
-      ])
